+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanelBuilder.java
@@ -20,10 +20,10 @@ package is.codion.swing.framework.ui;
 
 import is.codion.framework.db.EntityConnectionProvider;
 import is.codion.framework.domain.entity.EntityType;
+import is.codion.swing.common.ui.control.ControlIcon;
 
 import org.jspecify.annotations.Nullable;
 
-import javax.swing.ImageIcon;
 import java.util.Optional;
 import java.util.function.Function;
 
@@ -38,7 +38,7 @@ final class EntityPanelBuilder implements EntityPanel.Builder {
 
 	private @Nullable String caption;
 	private @Nullable String description;
-	private @Nullable ImageIcon icon;
+	private @Nullable ControlIcon icon;
 
 	EntityPanelBuilder(EntityType entityType, Function<EntityConnectionProvider, EntityPanel> entityPanel) {
 		this.entityType = requireNonNull(entityType);
@@ -73,13 +73,13 @@ final class EntityPanelBuilder implements EntityPanel.Builder {
 	}
 
 	@Override
-	public EntityPanel.Builder icon(@Nullable ImageIcon icon) {
+	public EntityPanel.Builder icon(@Nullable ControlIcon icon) {
 		this.icon = icon;
 		return this;
 	}
 
 	@Override
-	public Optional<ImageIcon> icon() {
+	public Optional<ControlIcon> icon() {
 		return Optional.ofNullable(icon);
 	}